refactor(dirs): renamed `src/../https-errors/base.ts` to `http-exception.ts`
