+++ src/modules/base/http-errors/forbidden.ts
@@ -1,6 +1,6 @@
 import type { HttpStatusNames } from '#types/api';
 
-import { I18nHttpException } from './base';
+import { I18nHttpException } from './http-exception';
 
 export class i18nForbiddenException extends I18nHttpException {
 	override statusCode: number = 403;
