+++ src/modules/base/http-errors/http-exception.ts
