+++ src/modules/base/http-errors/index.ts
@@ -2,4 +2,4 @@ export * from './forbidden';
 export * from './not-found';
 export * from './unauthorized';
 export * from './bad-request';
-export * from './base';
+export * from './http-exception';
