+++ src/modules/base/http-errors/unauthorized.ts
@@ -1,6 +1,6 @@
 import type { HttpStatusNames } from '#types/api';
 
-import { I18nHttpException } from './base';
+import { I18nHttpException } from './http-exception';
 
 export class i18nUnauthorizedException extends I18nHttpException {
 	override statusCode: number = 401;
